move trait to pipelines package
